refactor: wave5 peels for types tracking, merger parse, task API
Behavior-preserving domain extractions to shrink god-files:
- types/task-tracking.ts (PR/issue/GitHub/GitLab tracking contracts)
- merger-git-parse.ts (parseFailingFiles, parsePorcelainZ, parseShortstat)
- app/api/tasks.ts (task CRUD client) + proxyApi/withNodeId on client
Ratchet line-count baselines for merger/types/legacy. Fix api-tasks
batchUpdateTaskModels call site that omitted thinkingLevel before projectId.
