+++ scripts/line-count-baseline.json
@@ -16,8 +16,8 @@
   "packages/core/src/mission-store.ts": 4364,
   "packages/core/src/postgres/sqlite-migrator.ts": 2068,
   "packages/core/src/store.ts": 2631,
-  "packages/core/src/types.ts": 7336,
-  "packages/dashboard/app/api/legacy.ts": 11557,
+  "packages/core/src/types.ts": 7304,
+  "packages/dashboard/app/api/legacy.ts": 11169,
   "packages/dashboard/app/components/AgentDetailView.tsx": 5552,
   "packages/dashboard/app/components/AgentsView.tsx": 2254,
   "packages/dashboard/app/components/ChatView.tsx": 3850,
@@ -92,7 +92,7 @@
   "packages/engine/src/agent-tools.ts": 4793,
   "packages/engine/src/executor.ts": 19107,
   "packages/engine/src/merger-ai.ts": 2148,
-  "packages/engine/src/merger.ts": 11296,
+  "packages/engine/src/merger.ts": 11243,
   "packages/engine/src/pi.ts": 2799,
   "packages/engine/src/project-engine.ts": 5051,
   "packages/engine/src/scheduler.ts": 3370,
